requirements.txt
langchain
langchain-community
langchain-text-splitters
langchain-huggingface
langgraph
chromadb
sentence-transformers
python-dotenv
pypdf
tiktoken
requests
